[general] Created dummy platformconfig for use with latest ptxdist
I created a dummy platformconfig with mostly all values empty or
set to their defaults (e.g. paths). 

It allows to compile a toolchain with the latest versions of 
ptxdist. 

An empty file for platformconfig won't do the job as this causes
references to "/sysroot".

TODO: Find better solution. Or create handcrafted platformconfig with
      only those things defined required for an OSELAS chain. 



git-svn-id: https://svn.pengutronix.de/svn/oselas/toolchain/trunks/OSELAS.Toolchain-trunk@5661 f8d472c7-5700-0410-ac5a-87979cec3adf