[m68k] Updated project file for arch 'm68k' and GCC 4.3 RC 20080301
Added a config file using 
* binutils 2.18
* GCC 4.3 RC 20080301
* glibc 2.3.6  (uses new glibc source tarball layout)

for architecture M68K/mcfv4e. 

The otherwise available mcfv4e toolchains seem to be still broken. So
a _working_ toolchain for v4e (and the new v4m) targets, which we
can fix and recompile, is urgently needed.

Patches for glibc 2.3.6 and M68k can be derived from the ltib distro
as needed (that's why I didn't start with glibc 2.6.x or so).



git-svn-id: https://svn.pengutronix.de/svn/oselas/toolchain/trunks/OSELAS.Toolchain-trunk@5662 f8d472c7-5700-0410-ac5a-87979cec3adf